requirements.txt
pymupdf
faiss-cpu
openai
streamlit
sentence-transformers
python-dotenv
tiktoken
langchain
